refactor: package refactoring (structure)
