+++ README.md
@@ -76,7 +76,7 @@ abstract double getDistancePerLiter();
 ## 기능정의
 0. 렌트카 회사의 자동차 추가(초기화)
 1. 고객으로부터 목적지 이동거리 입력받음
-   - controller 를 통해 model, view 충돌 해결?
+   - controller 를 통해 model, oilInjection.view 충돌 해결?
 2. 이동거리를 통해 각 차량별 필요 연료 계산
 3. 차량 별로 필요한 연료를 주입
 4. 차량 별 주입 연료량 확인 보고서 출력
