+++ src/test/java/oilInjection/controller/OilInjectionControllerTest.java
@@ -1,18 +1,19 @@
 package oilInjection.controller;
 
-import oilInjection.RentCompany;
+import oilInjection.model.RentCompany;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class OilInjectionControllerTest {
 
     @Test
     void workingTest() {
         OilInjectionController oilInjectionController = new OilInjectionController();
-        oilInjectionController.getDistanceForEachCar(RentCompany.create());
+        RentCompany rentCompany = RentCompany.create();
+        oilInjectionController.getDistanceForEachCar(rentCompany);
+        oilInjectionController.makeOilReport(rentCompany);
     }
+
     @DisplayName("distance register test")
     @Test
     void distanceRegisterTest() {
