+++ src/test/java/oilInjection/model/CarTest.java
@@ -1,6 +1,5 @@
 package oilInjection.model;
 
-import oilInjection.RentCompany;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
